+++ Sources/SpotifyWebAPI/API/SpotifyAPI.swift
@@ -15,7 +15,7 @@ import OpenCombineFoundation
  The methods that require authorization scopes and/or an access token that was
  issued on behalf of a user are declared in conditional conformances where
  `AuthorizationManager` conforms to `SpotifyScopeAuthorizationManager`.
- This protcol requires conforming types to support authorization scopes.
+ This protocol requires conforming types to support authorization scopes.
  This strategy provides a compile-time guarantee that you cannot call methods
  that require authorization scopes if you are using an authorization manager
  that doesn't support them.
@@ -272,7 +272,7 @@ extension SpotifyAPI {
     /// Only use it for testing purposes.
     public func setupDebugging() {
         
-        // REMEMBER: keep up to date with "Debugging" wiki page
+        // MARK: keep up to date with "Debugging" wiki page
         // https://github.com/Peter-Schorn/SpotifyAPI/wiki/Debugging
         
         self.logger.logLevel = .trace