+++ Sources/SpotifyWebAPI/Object Model/Post Request Objects/Other Objects/TimeRange.swift
@@ -2,7 +2,7 @@ import Foundation
 
 /// A Time Range. Used by the `currentUserTopArtists(_:offset:limit:)`
 /// and `currentUserTopTracks(_:offset:limit:)` methods.
-public enum TimeRange: String, Codable {
+public enum TimeRange: String, Codable, Hashable, CaseIterable {
     
     /// Long term.
     case longTerm = "long_term"